% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_rec_catch.R
\name{calibrate_rec_catch}
\alias{calibrate_rec_catch}
\title{Calibrate Recreational Catch}
\usage{
calibrate_rec_catch(
  p_star_cod,
  p_star_had,
  select_mode,
  k,
  select_season,
  directed_trips_file_path,
  catch_draws_file_path,
  MRIP_comparison,
  pstar_file_path
)
}
\arguments{
\item{Inputs}{include p_star_cod (value between 0:1), p_star_had (value between 0:1), select_mode (Private ("pr"), Shore ("sh"), or For-Hire ("fh")),
k (numeric value in loop used to calculate all simulations), select_season (numeric value meaning indicating either open (1) or closed (0)),
directed_trips_file_path (file path character), catch_draws_file_path (file path character), MRIP_comparison (csv) , pstar_file_path (file path character)}
}
\description{
Calculate recreational catch to produce costs (costs_new_all) and periods (pds_new_all) files.
}
\examples{
library(ggplot2)
}
