% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rec_catch.R
\name{predict_rec_catch}
\alias{predict_rec_catch}
\title{Predict Recreational Catch}
\usage{
predict_rec_catch(
  x,
  directed_trips_table,
  calibration_data_table,
  costs_new_all,
  catch_data_all,
  discard_mortality_dat,
  n_drawz = 50,
  n_catch_draws = 30,
  eff_seed = 130
)
}
\arguments{
\item{Inputs}{include p_star_cod (value between 0:1), p_star_had (value between 0:1), select_mode (Private ("pr"), Shore ("sh"), or For-Hire ("fh")),
k (numeric value in loop used to calculate all simulations), select_season (numeric value meaning indicating either open (1) or closed (0)),
directed_trips_file_path (file path character), catch_draws_file_path (file path character), MRIP_comparison (csv) , pstar_file_path (file path character)}
}
\description{
Calculate recreational catch to produce prediction dataframe files.
}
\examples{
library(ggplot2)
}
