% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rec_catch.R
\name{predict_rec_catch}
\alias{predict_rec_catch}
\title{Predict Recreational Catch}
\usage{
predict_rec_catch(
  x,
  select_season,
  calibration_data_table,
  directed_trips_table,
  size_data_read,
  costs_new_all,
  catch_data_all,
  n_drawz = 50,
  n_catch_draws = 30,
  eff_seed = 130
)
}
\arguments{
\item{Inputs}{include p_star_cod (value between 0:1), p_star_had (value between 0:1), select_mode (Private ("pr"), Shore ("sh"), or For-Hire ("fh")),
k (numeric value in loop used to calculate all simulations), select_season (numeric value meaning indicating either open (1) or closed (0)),
directed_trips_file_path (file path character), catch_draws_file_path (file path character), MRIP_comparison (csv) , pstar_file_path (file path character)}
}
\value{
A data frame containing the estimates of harvest, discards, effort, and angler welfare for the regulations in the directed trips file.
}
\description{
Predict recreational catch for either status-quo or alternative estimates of harvest, discards, effort, and angler welfare.
}
